+++ src/main/java/gov/usgs/volcanoes/swarm/rsam/RsamViewerFrame.java
@@ -1,14 +1,19 @@
 package gov.usgs.volcanoes.swarm.rsam;
 
 import java.awt.BorderLayout;
+import java.awt.Graphics;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.FileOutputStream;
 import java.util.HashMap;
 import java.util.List;
 
 import javax.swing.BorderFactory;
 import javax.swing.Box;
 import javax.swing.JButton;
+import javax.swing.JFileChooser;
 import javax.swing.JInternalFrame;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
@@ -18,10 +23,14 @@ import javax.swing.border.LineBorder;
 import javax.swing.event.InternalFrameAdapter;
 import javax.swing.event.InternalFrameEvent;
 
+import gov.usgs.volcanoes.core.contrib.PngEncoder;
+import gov.usgs.volcanoes.core.contrib.PngEncoderB;
 import gov.usgs.volcanoes.core.data.RSAMData;
 import gov.usgs.volcanoes.core.time.J2kSec;
 import gov.usgs.volcanoes.core.util.UiUtils;
 import gov.usgs.volcanoes.swarm.Icons;
+import gov.usgs.volcanoes.swarm.Swarm;
+import gov.usgs.volcanoes.swarm.SwarmConfig;
 import gov.usgs.volcanoes.swarm.SwarmUtil;
 import gov.usgs.volcanoes.swarm.Throbber;
 import gov.usgs.volcanoes.swarm.chooser.DataChooser;
@@ -49,6 +58,7 @@ public class RsamViewerFrame extends JInternalFrame implements Runnable, Setting
   private Thread updateThread;
   private boolean run;
   private JToolBar toolBar;
+  private JButton captureButton;
 
   private RsamViewSettings settings;
   private RsamViewPanel viewPanel;
@@ -125,7 +135,12 @@ public class RsamViewerFrame extends JInternalFrame implements Runnable, Setting
         });
     toolBar.add(ratioButton);
 
-    toolBar.addSeparator();
+    toolBar.addSeparator();    
+    
+    captureButton = SwarmUtil.createToolBarButton(Icons.camera, "Save RSAM image (P)",
+        new CaptureActionListener());
+    UiUtils.mapKeyStrokeToButton(this, "P", "capture", captureButton);
+    toolBar.add(captureButton);
 
     toolBar.add(Box.createHorizontalGlue());
 
@@ -254,5 +269,51 @@ public class RsamViewerFrame extends JInternalFrame implements Runnable, Setting
 
     spanIndex = i;
     getRsam();
+  }
+  
+  class CaptureActionListener implements ActionListener {
+    public void actionPerformed(final ActionEvent e) {
+      final JFileChooser chooser = new JFileChooser();
+      final File lastPath = new File(SwarmConfig.getInstance().lastPath);
+      chooser.setCurrentDirectory(lastPath);
+      String filename = "rsam_" + channel.trim().replaceAll(" ", "_") + ".png";
+      chooser.setSelectedFile(new File(filename));
+      chooser.setDialogTitle("Save RSAM Screen Capture");
+      final int result = chooser.showSaveDialog(Swarm.getApplicationFrame());
+      File f = null;
+      if (result == JFileChooser.APPROVE_OPTION) {
+        f = chooser.getSelectedFile();
+
+        if (f.exists()) {
+          final int choice = JOptionPane.showConfirmDialog(Swarm.getApplicationFrame(),
+              "File exists, overwrite?", "Confirm", JOptionPane.YES_NO_OPTION);
+          if (choice != JOptionPane.YES_OPTION) {
+            return;
+          }
+        }
+        SwarmConfig.getInstance().lastPath = f.getParent();
+      }
+      if (f == null) {
+        return;
+      }
+
+      int height = viewPanel.getHeight();
+      int width = viewPanel.getWidth();
+      
+      final BufferedImage image = new BufferedImage(width, height, BufferedImage.TYPE_4BYTE_ABGR);
+      final Graphics g = image.getGraphics();
+      viewPanel.paint(g);
+      g.translate(0, height);
+      
+      try {
+        final PngEncoderB png = new PngEncoderB(image, false, PngEncoder.FILTER_NONE, 7);
+        final FileOutputStream out = new FileOutputStream(f);
+        final byte[] bytes = png.pngEncode();
+        out.write(bytes);
+        out.close();
+      } catch (final Exception ex) {
+        ex.printStackTrace();
+      }
+    }
   }
 }